# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API RESTFUL Laravel 5.8",
        "description": "API documentation project",
        "contact": {
            "email": "carlos.clayton@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Show a category",
                "description": "Return a category",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "description": "Update a category",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store categories"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "description": "Delete a category",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/categories/trashed": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List of trashed categories",
                "description": "Return a list of trashed categories",
                "operationId": "ApiVue\\Http\\Controllers\\CategoriesController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/categories/restore/{id}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Restore a category",
                "description": "Restore a category",
                "operationId": "getCategoryById",
                "responses": {
                    "200": {
                        "description": "Store categories"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/api/orders/trashed": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List of trashed orders",
                "description": "Return a list of trashed orders",
                "operationId": "ApiVue\\Http\\Controllers\\OrdersController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/api/products/trashed": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List of trashed products",
                "description": "Return a list of trashed products",
                "operationId": "ApiVue\\Http\\Controllers\\ProductsController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List of users",
                "description": "Return a list of users",
                "operationId": "ApiVue\\Http\\Controllers\\UsersController::index",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store a user",
                "description": "Return message",
                "operationId": "ApiVue\\Http\\Controllers\\UsersController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store categories"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show a user",
                "description": "Return a user",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "description": "Return message",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update user"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "description": "Delete a user",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/users/trashed": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List of trashed users",
                "description": "Return a list of users",
                "operationId": "ApiVue\\Http\\Controllers\\UsersController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/users/restore/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Restore a user",
                "description": "Restore a user",
                "operationId": "getCategoryById",
                "responses": {
                    "200": {
                        "description": "Restore a user"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "scheme": "http"
            }
        }
    }
}